feat: grid layout for flag impact metrics
